+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.2.2.0 (2015-11-10)
+
+- `U` and `Z` to parse `UTCTime` and `ZonedTime` compatibly
+- Orphans `FromJSON` for `Day` and `LocalTime`
+
 # 0.2.1.0 (2015-10-05) GHC 7.6 Support
 
 - No `SymTag` or `SingObject` support